remove note about untested js implementation
the app tests only verify that, when talking to an actual postgres
database, the magic skunk strings are correct. those strings are shared,
so the database interaction is tested in JS in one sense (barring any
really weird skunk regressions that somehow aren't caught upstream --
this is almost an impossibility I think). also, testing the JS app is
included in the ci-test command, it's just that there aren't currently
any tests that get executed as part of it. if there are in the future
though, they'll be tested, and invoking the test command even without
any tests still at least cross-compiles the sjs code with scala 2.13 and
scala 3.